delete unused variable and resources
