+++ app/src/main/java/com/example/radiologi/LoginActivity.java
@@ -16,8 +16,8 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
-import com.example.radiologi.home.DataAdminActivity;
-import com.example.radiologi.home.DataDokterActivity;
+import com.example.radiologi.admin.home.DataAdminActivity;
+import com.example.radiologi.dokter.home.DataDokterActivity;
 
 import org.json.JSONArray;
 import org.json.JSONException;
