+++ app/src/main/java/com/example/radiologi/MyFirebaseMessagingService.java
@@ -11,8 +11,8 @@ import android.net.Uri;
 import android.os.Build;
 import android.util.Log;
 
-import com.example.radiologi.home.DataAdminActivity;
-import com.example.radiologi.home.DataDokterActivity;
+import com.example.radiologi.admin.home.DataAdminActivity;
+import com.example.radiologi.dokter.home.DataDokterActivity;
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
