+++ app/src/main/java/com/example/radiologi/admin/home/fragments/DataAdminPagerAdapter.java
@@ -1,9 +1,12 @@
-package com.example.radiologi;
+package com.example.radiologi.admin.home.fragments;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentStatePagerAdapter;
+
+import com.example.radiologi.admin.home.fragments.DataAdminBaruFragment;
+import com.example.radiologi.admin.home.fragments.DataAdminDiagnosaFragment;
 
 public class DataAdminPagerAdapter extends FragmentStatePagerAdapter {
