+++ app/src/main/res/layout/activity_data_admin.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background"
-    tools:context=".home.DataAdminActivity">
+    tools:context=".admin.home.DataAdminActivity">
 
     <RelativeLayout
         android:id="@+id/bar"
