+++ app/src/main/res/layout/fragment_data_admin_baru.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DataAdminBaruFragment">
+    tools:context=".admin.home.fragments.DataAdminBaruFragment">
 
     <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
         android:id="@+id/swipe_admin_data_baru"
