+++ app/src/main/res/layout/fragment_data_admin_diagnosa.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DataAdminDiagnosaFragment">
+    tools:context=".admin.home.fragments.DataAdminDiagnosaFragment">
 
     <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
         android:id="@+id/swipe_admin_terdiagnosa"
